testing/mochitest/tests/test_bug696.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=696
-->
<head>
  <title>Test for Bug 696</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=696">Mozilla Bug 696</a>
<p id="display"></p>
<div id="content" style="display: none">
  <table><tr id="mytr"><td>Foo</td><td>Bar</td></tr></table>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 696 **/
var mytr = $("content").getElementsByTagName("TR")[0];
is(mytr.getAttribute("ID"),"mytr","TR tags expose their ID attribute");

</script>
</pre>
</body>
</html>
